+++ ToDo-Garden/ToDoGardenUI/Sources/ToDoGardenUIComponent/UITableView+Extension.swift
@@ -1,15 +1,5 @@
 import UIKit
 
-public protocol ReusableIdentifier: AnyObject {
-  static var identifier: String { get }
-}
-
-extension ReusableIdentifier {
-  public static var identifier: String {
-    String(describing: self)
-  }
-}
-
 extension UITableViewCell: ReusableIdentifier { }
 
 extension UITableView {
